[query/pcode] Move PCode to physical types package
Summary:
+ PCode.scala defines PValue, PSettable, and PCode.
+ Abstract P*{Code,Value} definitions are alongside the appropriate
physical type.
+ Concrete PSettable and PCode definitions/implementations are with the
concrete physical type.

PCode is cross cutting concern between PTypes and ir.Emit code
generation. We decided that it would be best to organize with PCode
alongside PType, rather than alongside Emit.
