+++ hail/src/main/scala/is/hail/expr/ir/functions/ArrayFunctions.scala
@@ -4,7 +4,7 @@ import is.hail.annotations.Region
 import is.hail.asm4s._
 import is.hail.expr.ir._
 import is.hail.expr.types.coerce
-import is.hail.expr.types.physical.{PArray, PFloat64, PType}
+import is.hail.expr.types.physical.{PArray, PCode, PFloat64, PType}
 import is.hail.expr.types.virtual._
 import is.hail.utils._
