+++ hail/src/main/scala/is/hail/expr/ir/functions/IntervalFunctions.scala
@@ -3,7 +3,7 @@ package is.hail.expr.ir.functions
 import is.hail.annotations.{CodeOrdering, Region, StagedRegionValueBuilder}
 import is.hail.asm4s.{Code, _}
 import is.hail.expr.ir._
-import is.hail.expr.types.physical.{PBoolean, PCanonicalInterval, PInterval, PType}
+import is.hail.expr.types.physical.{PBoolean, PCanonicalInterval, PCode, PInterval, PIntervalCode, PType}
 import is.hail.expr.types.virtual.{TArray, TBoolean, TInt32, TInterval}
 import is.hail.utils._
