+++ hail/src/main/scala/is/hail/expr/ir/functions/StringFunctions.scala
@@ -4,7 +4,7 @@ import is.hail.annotations.{Region, StagedRegionValueBuilder}
 import is.hail.asm4s
 import is.hail.asm4s._
 import is.hail.expr.ir._
-import is.hail.expr.types.physical.{PArray, PBoolean, PCanonicalArray, PCanonicalString, PInt32, PInt64, PSet, PString, PType}
+import is.hail.expr.types.physical._
 import is.hail.expr.types.virtual._
 import is.hail.utils._
 import java.util.Locale
