+++ hail/src/main/scala/is/hail/expr/types/physical/PCanonicalDict.scala
@@ -2,7 +2,6 @@ package is.hail.expr.types.physical
 
 import is.hail.annotations.Region
 import is.hail.asm4s.Code
-import is.hail.expr.ir.{PCanonicalIndexableCode, PCode}
 import is.hail.expr.types.virtual.{TArray, TDict, Type}
 
 final case class PCanonicalDict(keyType: PType, valueType: PType, required: Boolean = false) extends PDict with PArrayBackedContainer {
