+++ hail/src/main/scala/is/hail/expr/types/physical/PCanonicalLocus.scala
@@ -1,11 +1,9 @@
-
 package is.hail.expr.types.physical
-import is.hail.variant.ReferenceGenome
-import is.hail.annotations._
-import is.hail.asm4s.{Code, coerce}
-import is.hail.expr.ir.EmitMethodBuilder
-import is.hail.variant._
 
+import is.hail.annotations._
+import is.hail.asm4s._
+import is.hail.expr.ir.{EmitCodeBuilder, EmitMethodBuilder}
+import is.hail.variant._
 
 object PCanonicalLocus {
   def apply(rg: ReferenceGenome): PLocus = PCanonicalLocus(rg.broadcastRG)
@@ -99,4 +97,48 @@ final case class PCanonicalLocus(rgBc: BroadcastRG, required: Boolean = false) e
       }
     }
   }
+}
+
+object PCanonicalLocusSettable {
+  def apply(sb: SettableBuilder, pt: PLocus, name: String): PCanonicalLocusSettable = {
+    new PCanonicalLocusSettable(pt,
+      sb.newSettable[Long](s"${ name }_a"),
+      sb.newSettable[Long](s"${ name }_contig"),
+      sb.newSettable[Int](s"${ name }_position"))
+
+  }
+}
+
+class PCanonicalLocusSettable(
+  val pt: PLocus,
+  val a: Settable[Long],
+  _contig: Settable[Long],
+  val position: Settable[Int]
+) extends PLocusValue with PSettable {
+  def get = new PCanonicalLocusCode(pt, a)
+
+  def store(pc: PCode): Code[Unit] = {
+    Code(
+      a := pc.asInstanceOf[PCanonicalLocusCode].a,
+      _contig := pt.contig(a),
+      position := pt.position(a))
+  }
+
+  def contig(): PStringCode = new PCanonicalStringCode(pt.contigType, _contig)
+}
+
+class PCanonicalLocusCode(val pt: PLocus, val a: Code[Long]) extends PLocusCode {
+  def code: Code[_] = a
+
+  def memoize(cb: EmitCodeBuilder, name: String, sb: SettableBuilder): PLocusValue = {
+    val s = PCanonicalLocusSettable(sb, pt, name)
+    cb.assign(s, this)
+    s
+  }
+
+  def memoize(cb: EmitCodeBuilder, name: String): PLocusValue = memoize(cb, name, cb.localBuilder)
+
+  def memoizeField(cb: EmitCodeBuilder, name: String): PLocusValue = memoize(cb, name, cb.fieldBuilder)
+
+  def store(mb: EmitMethodBuilder[_], r: Value[Region], dst: Code[Long]): Code[Unit] = Region.storeAddress(dst, a)
 }
