+++ hail/src/main/scala/is/hail/expr/types/physical/PLocus.scala
@@ -1,9 +1,9 @@
 package is.hail.expr.types.physical
 
 import is.hail.annotations._
-import is.hail.asm4s.{Code, coerce}
+import is.hail.asm4s._
 import is.hail.backend.BroadcastValue
-import is.hail.expr.ir.EmitMethodBuilder
+import is.hail.expr.ir.{EmitCodeBuilder, EmitMethodBuilder}
 import is.hail.expr.types.virtual.TLocus
 import is.hail.utils._
 import is.hail.variant._
@@ -32,4 +32,18 @@ abstract class PLocus extends ComplexPType {
   def position(address: Code[Long]): Code[Int]
 
   def positionType: PInt32
+}
+
+abstract class PLocusValue extends PValue {
+  def contig(): PStringCode
+
+  def position(): Value[Int]
+}
+
+abstract class PLocusCode extends PCode {
+  def pt: PLocus
+
+  def memoize(cb: EmitCodeBuilder, name: String): PLocusValue
+
+  def memoizeField(cb: EmitCodeBuilder, name: String): PLocusValue
 }
