+++ hail/src/test/scala/is/hail/expr/ir/agg/StagedBlockLinkedListSuite.scala
@@ -3,7 +3,7 @@ package is.hail.expr.ir.agg
 import scala.collection.generic.Growable
 import is.hail.annotations.{Region, SafeRow, ScalaToRegionValue, StagedRegionValueBuilder}
 import is.hail.asm4s.Code
-import is.hail.expr.ir.{EmitFunctionBuilder, EmitRegion, EmitCode, PCode}
+import is.hail.expr.ir.{EmitFunctionBuilder, EmitRegion, EmitCode}
 import is.hail.expr.types.physical._
 import is.hail.utils._
 import org.scalatest.testng.TestNGSuite
